requirements.txt
# --- Core API stack (known-good combo) ---
fastapi==0.116.1
uvicorn[standard]==0.30.6
httpx==0.27.2

requests==2.32.3

# Pydantic v2 family (do NOT pin pydantic-core; let pip resolve a matching build)
pydantic==2.10.4
pydantic-settings==2.7.1

# Microsoft Graph auth + simple docx creation
msal==1.31.0
python-docx==1.1.2

# Google Workspace
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1

# LLM provider SDKs (pins prevent transitive surprises)
openai==1.50.2
anthropic==0.34.2
google-generativeai==0.7.2
mistralai==1.2.3

# Database + async driver
SQLAlchemy==2.0.43
asyncpg==0.30.0

# Queue (RabbitMQ)
aio-pika==9.4.3

# AWS (sync SDK only for now to avoid resolver conflicts)
boto3==1.35.36
botocore==1.35.36
# aioboto3==13.2.0   # <-- Omit for now; this version often conflicts with botocore 1.35.x

# Observability
structlog==25.4.0
prometheus-client==0.22.1
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0

# Auth / Security
python-jose[cryptography]==3.5.0

# Utilities
tiktoken==0.11.0
redis==5.1.0
PyYAML==6.0.2
python-dotenv==1.0.1

# Optional: Official Slack client (uncomment if your integration uses it)
# slack-sdk==3.33.4
